Add cells class, cell class  and logic module
Cells object contains a collection of 64 objects each representing a
square on the board with the personalized to_s method.

Cell object represents a single square on the board with the to_s method.
The properties for now are:
* last -> previous square coordinates
* coord -> square coordinates
* piece -> piece object in this square
The coordinates are stored as a hash. The key->value pair represents
human-readable coordinates -> array coordinates

Logic module will contain all the logic necessary that shouldn't be
in any class.