Commit Alpha version of the game
Huge commit, most of the code was rewritten
New files added:
cell.rb - a class representing a single square on the board. In charge
of storing the info about the piece in it, it's color, coord and human
name (e.g "E2"), also points to the squares around it.

init_config.rb - a module for pieces, symbols and directions initial
configs.

moves.rb - a module that generates moves for all the pieces

pieces.rb - class in charge of creating pieces. To be deleted in the
future